size_t -> unsigned for common

+++ src/amun/common/beam.cpp
@@ -6,14 +6,14 @@ using namespace std;
 namespace amunmt {
 
 
-std::string Debug(const Beam &vec, size_t verbosity)
+std::string Debug(const Beam &vec, unsigned verbosity)
 {
   std::stringstream strm;
 
   strm << "size=" << vec.size();
 
   if (verbosity) {
-    for (size_t i = 0; i < vec.size(); ++i) {
+    for (unsigned i = 0; i < vec.size(); ++i) {
       const HypothesisPtr &hypo = vec[i];
       strm << " " << hypo->GetWord();
     }
@@ -22,14 +22,14 @@ std::string Debug(const Beam &vec, size_t verbosity)
   return strm.str();
 }
 
-std::string Debug(const Beams &vec, size_t verbosity)
+std::string Debug(const Beams &vec, unsigned verbosity)
 {
   std::stringstream strm;
 
   strm << "size=" << vec.size();
 
   if (verbosity) {
-    for (size_t i = 0; i < vec.size(); ++i) {
+    for (unsigned i = 0; i < vec.size(); ++i) {
       const Beam &beam = vec[i];
       strm << endl << "\t" << Debug(beam, verbosity);
     }
