+++ src/amun/common/god.cpp
@@ -64,7 +64,7 @@ God& God::Init(int argc, char** argv) {
 
   if (Get("source-vocab").IsSequence()) {
     YAML::Node tabVocabs = Get("source-vocab");
-    for (size_t i = 0; i < tabVocabs.size(); i++) {
+    for (unsigned i = 0; i < tabVocabs.size(); i++) {
       if (tabVocabs[i].IsSequence()) {
         auto paths = tabVocabs[i].as<std::vector<std::string>>();
         sourceVocabs_.emplace_back(FactorVocab(paths));
@@ -94,7 +94,7 @@ God& God::Init(int argc, char** argv) {
   returnNBestList_ = Get<bool>("n-best");
   useFusedSoftmax_ = true;
   if (gpuLoaders_.size() != 1 || // more than 1 scorer
-      God::Get<size_t>("beam-size") > 11 // beam size affect shared mem alloc in gLogSoftMax()
+      God::Get<unsigned>("beam-size") > 11 // beam size affect shared mem alloc in gLogSoftMax()
       ) {
     useFusedSoftmax_ = false;
   }
@@ -111,7 +111,7 @@ God& God::Init(int argc, char** argv) {
 
   LoadPrePostProcessing();
 
-  size_t totalThreads = GetTotalThreads();
+  unsigned totalThreads = GetTotalThreads();
   LOG(info)->info("Total number of threads: {}", totalThreads);
   amunmt_UTIL_THROW_IF2(totalThreads == 0, "Total number of threads is 0");
 
@@ -131,8 +131,8 @@ void God::Cleanup()
 void God::LoadScorers() {
   LOG(info)->info("Loading scorers...");
 #ifdef CUDA
-  size_t gpuThreads = God::Get<size_t>("gpu-threads");
-  auto devices = God::Get<std::vector<size_t>>("devices");
+  unsigned gpuThreads = God::Get<unsigned>("gpu-threads");
+  auto devices = God::Get<std::vector<unsigned>>("devices");
   if (gpuThreads > 0 && devices.size() > 0) {
     for (auto&& pair : config_.Get()["scorers"]) {
       std::string name = pair.first.as<std::string>();
@@ -141,7 +141,7 @@ void God::LoadScorers() {
   }
 #endif
 #ifdef HAS_CPU
-  size_t cpuThreads = God::Get<size_t>("cpu-threads");
+  unsigned cpuThreads = God::Get<unsigned>("cpu-threads");
   if (cpuThreads) {
     for (auto&& pair : config_.Get()["scorers"]) {
       std::string name = pair.first.as<std::string>();
@@ -150,7 +150,7 @@ void God::LoadScorers() {
   }
 #endif
 #ifdef HAS_FPGA
-  size_t fpgaThreads = God::Get<size_t>("fpga-threads");
+  unsigned fpgaThreads = God::Get<unsigned>("fpga-threads");
   if (fpgaThreads) {
     for (auto&& pair : config_.Get()["scorers"]) {
       std::string name = pair.first.as<std::string>();
@@ -168,15 +168,15 @@ void God::LoadFiltering() {
     LOG(info)->info("Reading target softmax filter file from {}", alignmentFile);
     Filter* filter = nullptr;
     if (filterOptions.size() >= 3) {
-      const size_t numNFirst = stoi(filterOptions[1]);
-      const size_t maxNumTranslation = stoi(filterOptions[2]);
+      const unsigned numNFirst = stoi(filterOptions[1]);
+      const unsigned maxNumTranslation = stoi(filterOptions[2]);
       filter = new Filter(GetSourceVocab(0, 0),
                           GetTargetVocab(),
                           alignmentFile,
                           numNFirst,
                           maxNumTranslation);
     } else if (filterOptions.size() == 2) {
-      const size_t numNFirst = stoi(filterOptions[1]);
+      const unsigned numNFirst = stoi(filterOptions[1]);
       filter = new Filter(GetSourceVocab(0, 0),
                           GetTargetVocab(),
                           alignmentFile,
@@ -193,7 +193,7 @@ void God::LoadFiltering() {
 void God::LoadPrePostProcessing() {
   if (Has("bpe")) {
     if(Get("bpe").IsSequence()) {
-      size_t i = 0;
+      unsigned i = 0;
       for(auto bpePath : Get<std::vector<std::string>>("bpe")) {
         LOG(info)->info("using bpe: {}", bpePath);
         preprocessors_.push_back(std::vector<PreprocessorPtr>());
@@ -215,11 +215,11 @@ void God::LoadPrePostProcessing() {
   }
 }
 
-Vocab& God::GetSourceVocab(size_t tab, size_t factor) const {
+Vocab& God::GetSourceVocab(unsigned tab, unsigned factor) const {
   return sourceVocabs_[tab].GetVocab(factor);
 }
 
-FactorVocab& God::GetSourceVocabs(size_t tab) const {
+FactorVocab& God::GetSourceVocabs(unsigned tab) const {
   return sourceVocabs_[tab];
 }
 
@@ -293,7 +293,7 @@ const std::map<std::string, float>& God::GetScorerWeights() const {
 }
 
 std::vector<std::vector<std::string>> God::Preprocess (
-  size_t i, const std::vector<std::vector<std::string>>& input
+  unsigned i, const std::vector<std::vector<std::string>>& input
 ) const {
   std::vector<std::vector<std::string>> processed = input;
   if (preprocessors_.size() >= i + 1) {
@@ -304,7 +304,7 @@ std::vector<std::vector<std::string>> God::Preprocess (
   return processed;
 }
 
-std::vector<std::string> God::Preprocess(size_t i, const std::vector<std::string>& input) const {
+std::vector<std::string> God::Preprocess(unsigned i, const std::vector<std::string>& input) const {
   std::vector<std::string> processed = input;
   if (preprocessors_.size() >= i + 1) {
     for (const auto& processor : preprocessors_[i]) {
@@ -326,25 +326,25 @@ DeviceInfo God::GetNextDevice() const
 {
   DeviceInfo ret;
 
-  size_t cpuThreads = 0, gpuThreads = 0, fpgaThreads = 0;
-  std::vector<size_t> gpuDevices, fpgaDevices;
+  unsigned cpuThreads = 0, gpuThreads = 0, fpgaThreads = 0;
+  std::vector<unsigned> gpuDevices, fpgaDevices;
 
 #ifdef CUDA
-  gpuThreads = Get<size_t>("gpu-threads");
-  gpuDevices = Get<std::vector<size_t>>("devices");
+  gpuThreads = Get<unsigned>("gpu-threads");
+  gpuDevices = Get<std::vector<unsigned>>("devices");
 #endif
 
 #ifdef HAS_CPU
-  cpuThreads = God::Get<size_t>("cpu-threads");
+  cpuThreads = God::Get<unsigned>("cpu-threads");
 #endif
 
 #ifdef HAS_FPGA
-  fpgaThreads = Get<size_t>("fpga-threads");
-  fpgaDevices = Get<std::vector<size_t>>("fpga-devices");
+  fpgaThreads = Get<unsigned>("fpga-threads");
+  fpgaDevices = Get<std::vector<unsigned>>("fpga-devices");
 #endif
 
-  size_t totGPUThreads = gpuThreads * gpuDevices.size();
-  size_t totFPGAThreads = fpgaThreads * fpgaDevices.size();
+  unsigned totGPUThreads = gpuThreads * gpuDevices.size();
+  unsigned totFPGAThreads = fpgaThreads * fpgaDevices.size();
 
   // start locking
   boost::unique_lock<boost::shared_mutex> lock(accessLock_);
@@ -355,21 +355,21 @@ DeviceInfo God::GetNextDevice() const
   }
   else if (threadIncr_ < cpuThreads + totGPUThreads) {
     ret.deviceType = GPUDevice;
-    size_t threadIncr = threadIncr_ - cpuThreads;
+    unsigned threadIncr = threadIncr_ - cpuThreads;
 
     ret.threadInd = threadIncr / gpuDevices.size();
 
-    size_t deviceInd = threadIncr % gpuDevices.size();
+    unsigned deviceInd = threadIncr % gpuDevices.size();
     assert(deviceInd < gpuDevices.size());
     ret.deviceId = gpuDevices[deviceInd];
   }
   else if (threadIncr_ < cpuThreads + totGPUThreads + totFPGAThreads) {
     ret.deviceType = FPGADevice;
-    size_t threadIncr = threadIncr_ - cpuThreads - totGPUThreads;
+    unsigned threadIncr = threadIncr_ - cpuThreads - totGPUThreads;
 
     ret.threadInd = threadIncr / fpgaDevices.size();
 
-    size_t deviceInd = threadIncr % fpgaDevices.size();
+    unsigned deviceInd = threadIncr % fpgaDevices.size();
     assert(deviceInd < fpgaDevices.size());
     ret.deviceId = fpgaDevices[deviceInd];
   }
@@ -388,26 +388,26 @@ Search &God::GetSearch() const
   return obj;
 }
 
-size_t God::GetTotalThreads() const
+unsigned God::GetTotalThreads() const
 {
-  size_t totalThreads = 0;
+  unsigned totalThreads = 0;
 
 #ifdef CUDA
-  size_t gpuThreads = Get<size_t>("gpu-threads");
-  auto devices = Get<std::vector<size_t>>("devices");
+  unsigned gpuThreads = Get<unsigned>("gpu-threads");
+  auto devices = Get<std::vector<unsigned>>("devices");
   LOG(info)->info("Setting GPU thread count to {}", gpuThreads);
   totalThreads += gpuThreads * devices.size();
 #endif
 
 #ifdef HAS_CPU
-  size_t cpuThreads = Get<size_t>("cpu-threads");
+  unsigned cpuThreads = Get<unsigned>("cpu-threads");
   LOG(info->info("Setting CPU thread count to {}", cpuThreads));
   totalThreads += cpuThreads;
 #endif
 
 #ifdef HAS_FPGA
-  size_t fpgaThreads = Get<size_t>("fpga-threads");
-  auto fpgaDevices = Get<std::vector<size_t>>("fpga-devices");
+  unsigned fpgaThreads = Get<unsigned>("fpga-threads");
+  auto fpgaDevices = Get<std::vector<unsigned>>("fpga-devices");
   LOG(info->info("Setting FPGA thread count to {}", fpgaThreads));
   totalThreads += fpgaThreads * fpgaDevices.size();
 #endif
