+++ src/amun/common/histories.cpp
@@ -8,7 +8,7 @@ namespace amunmt {
 Histories::Histories(const Sentences& sentences, bool normalizeScore)
  : coll_(sentences.size())
 {
-  for (size_t i = 0; i < sentences.size(); ++i) {
+  for (unsigned i = 0; i < sentences.size(); ++i) {
     const Sentence &sentence = sentences.Get(i);
     History *history = new History(sentence, normalizeScore, 3 * sentence.size());
     coll_[i].reset(history);
@@ -34,7 +34,7 @@ void Histories::SortByLineNum()
 
 void Histories::Append(const Histories &other)
 {
-  for (size_t i = 0; i < other.size(); ++i) {
+  for (unsigned i = 0; i < other.size(); ++i) {
     std::shared_ptr<History> history = other.coll_[i];
     coll_.push_back(history);
   }
