+++ src/amun/common/printer.cpp
@@ -4,7 +4,7 @@ using namespace std;
 
 namespace amunmt {
 
-std::vector<size_t> GetAlignment(const HypothesisPtr& hypothesis) {
+std::vector<unsigned> GetAlignment(const HypothesisPtr& hypothesis) {
   std::vector<SoftAlignment> aligns;
   HypothesisPtr last = hypothesis->GetPrevHyp();
   while (last->GetPrevHyp().get() != nullptr) {
@@ -12,10 +12,10 @@ std::vector<size_t> GetAlignment(const HypothesisPtr& hypothesis) {
     last = last->GetPrevHyp();
   }
 
-  std::vector<size_t> alignment;
+  std::vector<unsigned> alignment;
   for (auto it = aligns.rbegin(); it != aligns.rend(); ++it) {
-    size_t maxArg = 0;
-    for (size_t i = 0; i < it->size(); ++i) {
+    unsigned maxArg = 0;
+    for (unsigned i = 0; i < it->size(); ++i) {
       if ((*it)[maxArg] < (*it)[i]) {
         maxArg = i;
       }
@@ -27,10 +27,10 @@ std::vector<size_t> GetAlignment(const HypothesisPtr& hypothesis) {
 }
 
 
-std::string GetAlignmentString(const std::vector<size_t>& alignment) {
+std::string GetAlignmentString(const std::vector<unsigned>& alignment) {
   std::stringstream alignString;
   alignString << " |||";
-  for (size_t wordIdx = 0; wordIdx < alignment.size(); ++wordIdx) {
+  for (unsigned wordIdx = 0; wordIdx < alignment.size(); ++wordIdx) {
     alignString << " " << wordIdx << "-" << alignment[wordIdx];
   }
   return alignString.str();
@@ -48,13 +48,13 @@ std::string GetSoftAlignmentString(const HypothesisPtr& hypothesis) {
   alignString << " |||";
   for (auto it = aligns.rbegin(); it != aligns.rend(); ++it) {
     alignString << " ";
-    for (size_t i = 0; i < it->size(); ++i) {
+    for (unsigned i = 0; i < it->size(); ++i) {
       if (i>0) alignString << ",";
       alignString << (*it)[i];
     }
     // alternate code: distribute probability mass from alignment to <eos>
     // float aligned_to_eos = (*it)[it->size()-1];
-    // for (size_t i = 0; i < it->size()-1; ++i) {
+    // for (unsigned i = 0; i < it->size()-1; ++i) {
     //  if (i>0) alignString << ",";
     //  alignString << ( (*it)[i] / (1-aligned_to_eos) );
     // }
@@ -63,7 +63,7 @@ std::string GetSoftAlignmentString(const HypothesisPtr& hypothesis) {
   return alignString.str();
 }
 
-std::string GetNematusAlignmentString(const HypothesisPtr& hypothesis, std::string best, std::string source, size_t linenum) {
+std::string GetNematusAlignmentString(const HypothesisPtr& hypothesis, std::string best, std::string source, unsigned linenum) {
   std::vector<SoftAlignment> aligns;
   HypothesisPtr last = hypothesis;
   while (last->GetPrevHyp().get() != nullptr) {
@@ -80,7 +80,7 @@ std::string GetNematusAlignmentString(const HypothesisPtr& hypothesis, std::stri
   std::stringstream alignString;
   for (auto it = aligns.rbegin(); it != aligns.rend(); ++it) {
     alignString << "\n";
-    for (size_t i = 0; i < srcspaces+2; ++i) {
+    for (unsigned i = 0; i < srcspaces+2; ++i) {
       if (i>0) alignString << " ";
       alignString << (*it)[i];
     }
